+++ Insurance-System/src/main/java/org/example/business/education/view/EducationView.java
@@ -2,6 +2,8 @@ package org.example.business.education.view;
 
 import org.example.common.DepartmentView;
 import org.example.common.dto.RequestDto;
+import org.example.common.dto.ResponseDto;
+import org.example.planTeam.Status;
 
 public class EducationView extends DepartmentView {
