+++ Insurance-System/src/main/java/org/example/planTeam/design/view/DesignTeamView.java
@@ -1,6 +1,6 @@
 package org.example.planTeam.design.view;
 
-import org.example.common.view.DepartmentView;
+import org.example.common.DepartmentView;
 import org.example.common.dto.RequestDto;
 import org.example.common.dto.ResponseDto;
 import org.example.planTeam.Status;
